+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.5-dev.2 - 2020-10-07
+
+- L10n: Fixed locale prefix format for Android in `from_xml`.
+
+## 0.2.5-dev.1 - 2020-10-06
+
+- L10n: `from_xml` supports `ru_RU` format for locales.
+- L10n: generation in release mode.
+- Pass `--verbose` argument in `pub run` command, if verbose enabled.
+
 ## 0.2.4-dev.1 - 2020-09-16
 
 - L10n: handled Android restricted locale names.